pyproject.toml
[tool.poetry]
name = "heathrow-noise"
version = "0.1.0"
description = "Heathrow runway alternation tracker — publishes noise impact sensors to Home Assistant via MQTT"
authors = ["Ron Schaeffer <ron@ronschaeffer.com>"]
readme = "README.md"
packages = [{include = "heathrow_noise", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.11"
ha-mqtt-publisher = ">=0.4.0"
fastapi = ">=0.110.0"
uvicorn = {extras = ["standard"], version = ">=0.29.0"}
httpx = ">=0.27.0"
pyyaml = ">=6.0"
haversine = ">=2.8.0"
feedparser = ">=6.0.11"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0"
pytest-cov = ">=5.0"
ruff = ">=0.4.0"
pre-commit = ">=3.7"

[tool.poetry.scripts]
heathrow-noise = "heathrow_noise.__main__:main"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "G004", "UP"]

[tool.ruff.lint.isort]
force-sort-within-sections = true

[tool.ruff.format]
quote-style = "double"
line-ending = "lf"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
